Added trade, added submodules
