--- cfxws/__init__.py
@@ -1,6 +1,0 @@
-import exchange
-import client
-
-exchanges = {
-    "binance": Binance
-}
