--- cfxws/fromclass.py
@@ -1,8 +1,0 @@
-from exchange import Binance
-
-b = Binance()
-
-def echo(self, data):
-    print(data)
-
-b.listen_ticks(echo, ['adabtc', 'ethbtc'])
