--- cfxws/requirements.txt
@@ -1,44 +1,0 @@
-aiodns==1.1.1
-aiohttp==2.3.7
-amqp==2.2.2
-asn1crypto==0.24.0
-async-timeout==2.0.0
-asyncio==3.4.3
-attrs==17.4.0
-autobahn==17.10.1
-Automat==0.6.0
-billiard==3.5.0.3
-cchardet==2.1.1
-cctx==1.0.0
-ccxt==1.10.621
-celery==4.1.0
-certifi==2017.11.5
-cffi==1.11.4
-chardet==3.0.4
-constantly==15.1.0
-cryptography==2.1.4
-hyperlink==17.3.1
-idna==2.6
-incremental==17.5.0
-influxdb==5.0.0
-kombu==4.1.0
-multidict==3.3.2
-pyasn1==0.4.2
-pyasn1-modules==0.2.1
-pycares==2.3.0
-pycparser==2.18
-pyOpenSSL==17.5.0
-python-dateutil==2.6.1
-pytz==2017.3
-raven==6.4.0
-redis==2.10.6
-redispy==3.0.0
-requests==2.18.4
-service-identity==17.0.0
-six==1.11.0
-Twisted==17.9.0
-txaio==2.8.2
-urllib3==1.22
-vine==1.1.4
-yarl==0.17.0
-zope.interface==4.4.3
