+++ requirements.txt
@@ -1,0 +1,23 @@
+asn1crypto==0.24.0
+attrs==17.4.0
+autobahn==17.10.1
+Automat==0.6.0
+ccxt==1.10.849
+certifi==2018.1.18
+cffi==1.11.4
+chardet==3.0.4
+constantly==15.1.0
+cryptography==2.1.4
+enum34==1.1.6
+hyperlink==17.3.1
+idna==2.6
+incremental==17.5.0
+ipaddress==1.0.19
+pycparser==2.18
+pyOpenSSL==17.5.0
+requests==2.18.4
+six==1.11.0
+Twisted==17.9.0
+txaio==2.8.2
+urllib3==1.22
+zope.interface==4.4.3
